moved controller generator
