+++ _vendor/knplabs/rad-bundle/Knp/Bundle/RadBundle/Generator/ControllerGenerator.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Knp\Bundle\SoRBundle\Generator;
+namespace Knp\Bundle\RadBundle\Generator;
 
 use Symfony\Component\Filesystem\Filesystem;
 use Symfony\Component\HttpKernel\Bundle\BundleInterface;
@@ -17,7 +17,6 @@ class ControllerGenerator extends Generator
     private $filesystem;
     private $skeletonDir;
     private $bundle;
-    private $format;
 
     /**
      * Constructor.
@@ -37,7 +36,6 @@ class ControllerGenerator extends Generator
      * @param BundleInterface $bundle A bundle object
      * @param string $entity The entity relative class name
      * @param ClassMetadataInfo $metadata The entity class metadata
-     * @param string $format The configuration format (xml, yaml, annotation)
      * @param string $routePrefix The route name prefix
      * @param array $needWriteActions Wether or not to generate write actions
      *
@@ -45,58 +43,11 @@ class ControllerGenerator extends Generator
      */
     public function generate(BundleInterface $bundle, $controller)
     {
-        $this->bundle   = $bundle;
-        $this->setFormat('yml');
+        $this->bundle = $bundle;
 
         $this->generateControllerClass($controller);
 
         $dir = sprintf('%s/Resources/views/%s', $this->bundle->getPath(), str_replace('\\', '/', $controller));
-    }
-
-    /**
-     * Sets the configuration format.
-     *
-     * @param string $format The configuration format
-     */
-    private function setFormat($format)
-    {
-        switch ($format) {
-            case 'yml':
-            case 'xml':
-            case 'php':
-            case 'annotation':
-                $this->format = $format;
-                break;
-            default:
-                $this->format = 'yml';
-                break;
-        }
-    }
-
-    /**
-     * Generates the routing configuration.
-     *
-     */
-    private function generateConfiguration()
-    {
-        if (!in_array($this->format, array('yml', 'xml', 'php'))) {
-            return;
-        }
-
-        $target = sprintf(
-            '%s/Resources/config/routing/%s.%s',
-            $this->bundle->getPath(),
-            strtolower(str_replace('\\', '_', $this->entity)),
-            $this->format
-        );
-
-        $this->renderFile($this->skeletonDir, 'config/routing.'.$this->format, $target, array(
-            'actions'           => $this->actions,
-            'route_prefix'      => $this->routePrefix,
-            'route_name_prefix' => $this->routeNamePrefix,
-            'bundle'            => $this->bundle->getName(),
-            'entity'            => $this->entity,
-        ));
     }
 
     /**
@@ -122,7 +73,6 @@ class ControllerGenerator extends Generator
             'controller'        => $controller,
             'bundle'            => $this->bundle->getName(),
             'namespace'         => $this->bundle->getNamespace(),
-            'format'            => $this->format,
         ));
     }
