--- src/Knp/Bundle/SoRBundle/Controller/TestController.php
@@ -1,26 +1,0 @@
-<?php
-
-namespace Knp\Bundle\SoRBundle\Controller;
-
-use Symfony\Bundle\FrameworkBundle\Controller\Controller;
-
-/**
- * Test controller.
- *
- */
-class TestController extends Controller
-{
-    public function test()
-    {
-        // code...
-    }
-
-    /**
-    * test2 method.
-     *
-     */
-    public function test2Action()
-    {
-        return array();
-    }
- }
