--- src/Knp/Bundle/SoRBundle/KnpSoRBundle.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Knp\Bundle\SoRBundle;
-
-use Symfony\Component\HttpKernel\Bundle\Bundle;
-
-class KnpSoRBundle extends Bundle
-{
-}
