--- src/Knp/Bundle/SoRBundle/Resources/config/routing.yml
@@ -1,3 +1,0 @@
-KnpSoRBundle_homepage:
-    pattern:  /hello/{name}
-    defaults: { _controller: KnpSoRBundle:Default:index }
