--- src/Knp/Bundle/SoRBundle/Resources/config/services.yml
@@ -1,7 +1,0 @@
-parameters:
-#    knp_so_r.example.class: Knp\Bundle\SoRBundle\Example
-
-services:
-#    knp_so_r.example:
-#        class: %knp_so_r.example.class%
-#        arguments: [@service_id, "plain_value", %parameter%]
