--- src/Knp/Bundle/SoRBundle/Resources/skeleton/bundle/Bundle.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace {{ namespace }};
-
-use Symfony\Component\HttpKernel\Bundle\Bundle;
-
-class {{ bundle }} extends Bundle
-{
-}
