--- src/Knp/Bundle/SoRBundle/Resources/skeleton/bundle/index.html.twig
@@ -1,1 +1,0 @@
-Hello {% raw %}{{ name }}{% endraw %}!
