--- src/Knp/Bundle/SoRBundle/Resources/skeleton/bundle/routing.php
@@ -1,11 +1,0 @@
-<?php
-
-use Symfony\Component\Routing\RouteCollection;
-use Symfony\Component\Routing\Route;
-
-$collection = new RouteCollection();
-$collection->add('{{ bundle }}_homepage', new Route('/hello/{name}', array(
-    '_controller' => '{{ bundle }}:Default:index',
-)));
-
-return $collection;
