--- src/Knp/Bundle/SoRBundle/Resources/skeleton/bundle/routing.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-
-<routes xmlns="http://symfony.com/schema/routing"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://symfony.com/schema/routing http://symfony.com/schema/routing/routing-1.0.xsd">
-
-    <route id="{{ bundle }}_homepage" pattern="/hello/{name}">
-        <default key="_controller">{{ bundle }}:Default:index</default>
-    </route>
-</routes>
