--- src/Knp/Bundle/SoRBundle/Resources/skeleton/bundle/routing.yml
@@ -1,3 +1,0 @@
-{{ bundle }}_homepage:
-    pattern:  /hello/{name}
-    defaults: { _controller: {{ bundle }}:Default:index }
