--- src/Knp/Bundle/SoRBundle/Resources/skeleton/bundle/services.php
@@ -1,21 +1,0 @@
-<?php
-
-use Symfony\Component\DependencyInjection\Definition;
-use Symfony\Component\DependencyInjection\Reference;
-use Symfony\Component\DependencyInjection\Parameter;
-
-/*
-
-$container->setDefinition(
-    '{{ extension_alias }}.example',
-    new Definition(
-        '{{ namespace }}\Example',
-        array(
-            new Reference('service_id'),
-            "plain_value",
-            new Parameter('parameter_name'),
-        )
-    )
-);
-
-*/
