--- src/Knp/Bundle/SoRBundle/Resources/skeleton/bundle/services.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" ?>
-
-<container xmlns="http://symfony.com/schema/dic/services"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
-
-    <!--
-    <parameters>
-        <parameter key="{{ extension_alias }}.example.class">{{ namespace }}\Example</parameter>
-    </parameters>
-
-    <services>
-        <service id="{{ extension_alias }}.example" class="%{{ extension_alias }}.example.class%">
-            <argument type="service" id="service_id" />
-            <argument>plain_value</argument>
-            <argument>%parameter_name%</argument>
-        </service>
-    </services>
-    -->
-</container>
