--- src/Knp/Bundle/SoRBundle/Resources/skeleton/controller/action.php
@@ -1,13 +1,0 @@
-
-    /**
-    * {{ action }} method.
-     *
-{% if 'annotation' == format %}
-     * @Route("/new", name="{{ route_name_prefix }}_new")
-     * @Template()
-{% endif %}
-     */
-    public function {{ action }}Action()
-    {
-        return array();
-    }
