--- src/Knp/Bundle/SoRBundle/Resources/skeleton/crud/actions/delete.php
@@ -1,38 +1,0 @@
-
-    /**
-     * Deletes a {{ entity }} entity.
-     *
-{% if 'annotation' == format %}
-     * @Route("/{id}/delete", name="{{ route_name_prefix }}_delete")
-     * @Method("post")
-{% endif %}
-     */
-    public function deleteAction($id)
-    {
-        $form = $this->createDeleteForm($id);
-        $request = $this->getRequest();
-
-        $form->bindRequest($request);
-
-        if ($form->isValid()) {
-            $em = $this->getDoctrine()->getEntityManager();
-            $entity = $em->getRepository('{{ bundle }}:{{ entity }}')->find($id);
-
-            if (!$entity) {
-                throw $this->createNotFoundException('Unable to find {{ entity }} entity.');
-            }
-
-            $em->remove($entity);
-            $em->flush();
-        }
-
-        return $this->redirect($this->generateUrl('{{ route_name_prefix }}'));
-    }
-
-    private function createDeleteForm($id)
-    {
-        return $this->createFormBuilder(array('id' => $id))
-            ->add('id', 'hidden')
-            ->getForm()
-        ;
-    }
