--- src/Knp/Bundle/SoRBundle/Resources/skeleton/crud/actions/edit.php
@@ -1,36 +1,0 @@
-
-    /**
-     * Displays a form to edit an existing {{ entity }} entity.
-     *
-{% if 'annotation' == format %}
-     * @Route("/{id}/edit", name="{{ route_name_prefix }}_edit")
-     * @Template()
-{% endif %}
-     */
-    public function editAction($id)
-    {
-        $em = $this->getDoctrine()->getEntityManager();
-
-        $entity = $em->getRepository('{{ bundle }}:{{ entity }}')->find($id);
-
-        if (!$entity) {
-            throw $this->createNotFoundException('Unable to find {{ entity }} entity.');
-        }
-
-        $editForm = $this->createForm(new {{ entity_class }}Type(), $entity);
-        $deleteForm = $this->createDeleteForm($id);
-
-{% if 'annotation' == format %}
-        return array(
-            'entity'      => $entity,
-            'edit_form'   => $editForm->createView(),
-            'delete_form' => $deleteForm->createView(),
-        );
-{% else %}
-        return $this->render('{{ bundle }}:{{ entity|replace({'\\': '/'}) }}:edit.html.twig', array(
-            'entity'      => $entity,
-            'edit_form'   => $editForm->createView(),
-            'delete_form' => $deleteForm->createView(),
-        ));
-{% endif %}
-    }
