--- src/Knp/Bundle/SoRBundle/Resources/skeleton/crud/actions/update.php
@@ -1,48 +1,0 @@
-
-    /**
-     * Edits an existing {{ entity }} entity.
-     *
-{% if 'annotation' == format %}
-     * @Route("/{id}/update", name="{{ route_name_prefix }}_update")
-     * @Method("post")
-     * @Template("{{ bundle }}:{{ entity }}:edit.html.twig")
-{% endif %}
-     */
-    public function updateAction($id)
-    {
-        $em = $this->getDoctrine()->getEntityManager();
-
-        $entity = $em->getRepository('{{ bundle }}:{{ entity }}')->find($id);
-
-        if (!$entity) {
-            throw $this->createNotFoundException('Unable to find {{ entity }} entity.');
-        }
-
-        $editForm   = $this->createForm(new {{ entity_class }}Type(), $entity);
-        $deleteForm = $this->createDeleteForm($id);
-
-        $request = $this->getRequest();
-
-        $editForm->bindRequest($request);
-
-        if ($editForm->isValid()) {
-            $em->persist($entity);
-            $em->flush();
-
-            return $this->redirect($this->generateUrl('{{ route_name_prefix }}_edit', array('id' => $id)));
-        }
-
-{% if 'annotation' == format %}
-        return array(
-            'entity'      => $entity,
-            'edit_form'   => $editForm->createView(),
-            'delete_form' => $deleteForm->createView(),
-        );
-{% else %}
-        return $this->render('{{ bundle }}:{{ entity|replace({'\\': '/'}) }}:edit.html.twig', array(
-            'entity'      => $entity,
-            'edit_form'   => $editForm->createView(),
-            'delete_form' => $deleteForm->createView(),
-        ));
-{% endif %}
-    }
