--- src/Knp/Bundle/SoRBundle/Resources/skeleton/crud/controller.php
@@ -1,26 +1,0 @@
-<?php
-
-namespace {{ namespace }}\Controller;
-
-use Symfony\Bundle\FrameworkBundle\Controller\Controller;
-{% if 'annotation' == format -%}
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\Method;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\Template;
-{%- endif %}
-
-use {{ namespace }}\Entity\{{ entity }};
-{% if 'new' in actions or 'edit' in actions %}
-use {{ namespace }}\Form\{{ entity }}Type;
-{% endif %}
-
-/**
- * {{ entity }} controller.
- *
-{% if 'annotation' == format %}
- * @Route("/{{ route_prefix }}")
-{% endif %}
- */
-class {{ entity_class }}Controller extends Controller
-{
-}
