--- src/Knp/Bundle/SoRBundle/Resources/skeleton/crud/views/edit.html.twig
@@ -1,11 +1,0 @@
-<h1>{{ entity }} edit</h1>
-
-<form action="{{ "{{ path('"~ route_name_prefix ~"_update', { 'id': entity.id }) }}" }}" method="post" {{ "{{ form_enctype(edit_form) }}" }}>
-    {{ "{{ form_widget(edit_form) }}" }}
-    <p>
-        <button type="submit">Edit</button>
-    </p>
-</form>
-
-{% set hide_edit, hide_delete = true, false %}
-{% include 'views/others/record_actions.html.twig' %}
