--- src/Knp/Bundle/SoRBundle/Resources/skeleton/crud/views/others/actions.html.twig
@@ -1,12 +1,0 @@
-
-                <ul>
-
-                {%- for action in record_actions %}
-
-                    <li>
-                        <a href="{{ "{{ path('"~ route_name_prefix ~"_"~ action ~"', { 'id': entity.id }) }}" }}">{{ action }}</a>
-                    </li>
-
-                {%- endfor %}
-
-                </ul>
