--- src/Knp/Bundle/SoRBundle/Resources/skeleton/crud/views/others/record_actions.html.twig
@@ -1,22 +1,0 @@
-<ul class="record_actions">
-    <li>
-        <a href="{{ "{{ path('"~ route_name_prefix ~"') }}" }}">
-            Back to the list
-        </a>
-    </li>
-{% if ('edit' in actions) and (not hide_edit) %}
-    <li>
-        <a href="{{ "{{ path('"~ route_name_prefix ~"_edit', { 'id': entity.id }) }}" }}">
-            Edit
-        </a>
-    </li>
-{% endif %}
-{% if ('delete' in actions) and (not hide_delete) %}
-    <li>
-        <form action="{{ "{{ path('"~ route_name_prefix ~"_delete', { 'id': entity.id }) }}" }}" method="post">
-            {{ '{{ form_widget(delete_form) }}' }}
-            <button type="submit">Delete</button>
-        </form>
-    </li>
-{% endif %}
-</ul>
