--- src/Knp/Bundle/SoRBundle/Resources/skeleton/crud/views/show.html.twig
@@ -1,28 +1,0 @@
-<h1>{{ entity }}</h1>
-
-<table class="record_properties">
-    <tbody>
-    {%- for field, metadata in fields %}
-
-        <tr>
-            <th>{{ field|capitalize }}</th>
-
-        {%- if metadata.type in ['date', 'datetime'] %}
-
-            <td>{{ '{{ entity.'~ field|replace({'_': ''}) ~'|date(\'Y-m-d H:i:s\') }}' }}</td>
-
-        {%- else %}
-
-            <td>{{ '{{ entity.'~ field|replace({'_': ''}) ~' }}' }}</td>
-
-        {%- endif %}
-
-        </tr>
-
-    {%- endfor %}
-
-    </tbody>
-</table>
-
-{% set hide_edit, hide_delete = false, false %}
-{% include 'views/others/record_actions.html.twig' %}
