--- src/Knp/KnpSoR2Bundle.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Knp;
-
-use Symfony\Component\HttpKernel\Bundle\Bundle;
-
-class KnpSoR2Bundle extends Bundle
-{
-}
