--- src/Knp/Resources/config/services.yml
@@ -1,7 +1,0 @@
-parameters:
-#    knp_so_r2.example.class: Knp\Example
-
-services:
-#    knp_so_r2.example:
-#        class: %knp_so_r2.example.class%
-#        arguments: [@service_id, "plain_value", %parameter%]
